Revert "userspace, libc: try to use stdint types"
This reverts commit 37c091cf20ebf6a0fd516aa8e026af1d5af30382.

+++ pkgs/gfx/catimg.c
@@ -1,4 +1,3 @@
-#include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
@@ -45,7 +44,7 @@ int main(int argc, char **argv) {
     }
     struct fbdev_bitblit bitblit = {
         .target_buffer = GFX_FRONT_BUFFER,
-        .source = (uint32_t *)data,
+        .source = (unsigned long*)data,
         .x = 0,
         .y = 0,
         .width = w,
